# src/main/resources/api.yml
...
swagger: '2.0'
...
info:
  version: "0.0.1"
  title: Business Process API

# Describe your paths here
paths:
...
  /content:
    get:
      operationId: getProcessDefinitions
      consumes:
      - application/json
      parameters:
      - name: Authorization
        in: header
        type: string
        required: true
      produces:
      - application/json
      responses:
        # Response code
        200:
          description: Successful response
          schema:
            type: array
            items:
              $ref: '#/definitions/Process'
    post:
      operationId: addProcessDefinition
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/Process'
      - name: Authorization
        in: header
        type: string
        required: true

      responses:
        # Response code
        200:
          description: BP added
          schema:
            $ref: '#/definitions/ApiResponse'
    put:
      operationId: updateProcessDefinition
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: body
        schema:
          $ref: '#/definitions/Process'
      - name: Authorization
        in: header
        type: string
        required: true
      responses:
        # Response code
        200:
          description: BP updated
          schema:
            $ref: '#/definitions/ApiResponse'
  /content/{processID}:
    get:
      operationId: getProcessDefinition
      produces:
      - application/json
      parameters:
      - in: path
        name: processID
        type: string
        description: The id of the business process.
        required: true
      - name: Authorization
        in: header
        type: string
        required: true
      responses:
        # Response code
        200:
          description: BP obtained
          schema:
            $ref: '#/definitions/Process'
    delete:
      operationId: deleteProcessDefinition
      produces:
      - application/json
      parameters:
      - in: path
        name: processID
        type: string
        required: true
      - name: Authorization
        in: header
        type: string
        required: true
      responses:
        # Response code
        200:
          description: BP deleted
          schema:
            $ref: '#/definitions/ApiResponse'
...
  /application:
    post:
      operationId: addProcessConfiguration
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ProcessConfiguration'
      - name: Authorization
        in: header
        type: string
        required: true
      responses:
        # Response code
        200:
          description: Partner Preference added
          schema:
            $ref: '#/definitions/ApiResponse'
    put:
      operationId: updateProcessConfiguration
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: body
        schema:
          $ref: '#/definitions/ProcessConfiguration'
      - name: Authorization
        in: header
        type: string
        required: true
      responses:
        # Response code
        200:
          description: Partner Preference updated
          schema:
            $ref: '#/definitions/ApiResponse'
  /application/{partnerID}:
    get:
      operationId: getProcessConfiguration
      produces:
      - application/json
      parameters:
      - in: path
        name: partnerID
        type: string
        description: The id of the partner.
        required: true
      - name: Authorization
        in: header
        type: string
        required: true
      responses:
        # Response code
        200:
          description: Partner Preference obtained
          schema:
            type: array
            items:
              $ref: '#/definitions/ProcessConfiguration'
  /application/{partnerID}/{processID}/{roleType}:
    get:
      operationId: getProcessConfigurationByProcessID
      produces:
      - application/json
      parameters:
      - in: path
        name: partnerID
        type: string
        description: The id of the partner.
        required: true
      - in: path
        name: processID
        type: string
        required: true
      - in: path
        name: roleType
        type: string
        required: true
      - name: Authorization
        in: header
        type: string
        required: true
      responses:
        # Response code
        200:
          description: Partner Preference obtained
          schema:
            $ref: '#/definitions/ProcessConfiguration'
    delete:
      operationId: deleteProcessConfiguration
      produces:
      - application/json
      parameters:
      - in: path
        name: partnerID
        type: string
        required: true
      - in: path
        name: processID
        type: string
        required: true
      - in: path
        name: roleType
        type: string
        required: true
      - name: Authorization
        in: header
        type: string
        required: true        
      responses:
        # Response code
        200:
          description: Partner Preference deleted
          schema:
            $ref: '#/definitions/ApiResponse'
...
definitions:
...
  Transaction:
    title: The transaction in a business process
    type: object
    properties:
      transactionID:
        type: string
      initiatorRole:
        type: string
        enum:
        - BUYER
        - SELLER
        - LOGISTICSPROVIDER
      responderRole:
        type: string
        enum:
        - BUYER
        - SELLER
        - LOGISTICSPROVIDER
      documentType:
        type: string
        enum:
        - CATALOGUE
        - REQUESTFORQUOTATION
        - QUOTATION
        - ORDER
        - ORDERRESPONSESIMPLE
        - REMITTANCEADVICE
        - INVOICE
        - APPLICATIONRESPONSE
        - TRANSPORTATIONSTATUS
        - DESPATCHADVICE
        - RECEIPTADVICE
        - PPAPREQUEST
        - PPAPRESPONSE
        - TRANSPORTEXECUTIONPLANREQUEST
        - TRANSPORTEXECUTIONPLAN
        - ITEMINFORMATIONREQUEST
        - ITEMINFORMATIONRESPONSE
        - OTHER
  Process:
    title: Business Process Object
    type: object
    properties:
      processID:
        type: string
      processName:
        type: string
      processType:
        type: string
        enum:
        - CATALOGUE
        - NEGOTIATION
        - ORDER
        - REMITTANCEADVICE
        - INVOICE
        - TRACKING
        - FULFILMENT
        - PRODUCTCONFIGURATION
        - PPAP
        - TRANSPORT_EXECUTION_PLAN
        - ITEM_INFORMATION_REQUEST
        - OTHER
      textContent:
        type: string
      bpmnContent:
        type: string
      transactions:
        type: array
        items:
          $ref: '#/definitions/Transaction'
  ExecutionConfiguration:
    title: Execution Configuration Object
    type: object
    properties:
      applicationType:
        type: string
        enum:
        - DATAADAPTER
        - DATACHANNEL
        - DATAPROCESSOR
      executionType:
        type: string
        enum:
        - JAVA
        - MICROSERVICE
        - OTHER
      executionUri:
        type: string
  TransactionConfiguration:
    title: Application Configuration Object for each activity in a business process
    type: object
    properties:
      transactionID:
        type: string
        description: The transaction where the activity exist
      executionConfigurations:
        type: array
        items:
          $ref: '#/definitions/ExecutionConfiguration'
  ProcessConfiguration:
    title: Applications in Business Process Configuration Object
    type: object
    properties:
      partnerID:
        type: string
      roleType:
        type: string
        enum:
        - BUYER
        - SELLER
        - LOGISTICSPROVIDER        
      processID:
        type: string
      transactionConfigurations:
        type: array
        items:
          $ref: '#/definitions/TransactionConfiguration'
...
  ApiResponse:
    type: object
    properties:
      code:
        type: integer
      type:
        type: string
      message:
        type: string